Windows: Fix sp_[blocking|nonblocking]_read failure.
On Windows systems sp_*_read may fail spuriously with SP_ERR_FAIL and an
operating system error code ERROR_OPERATION_ABORTED (995, 0x3E3) when the
last call to sp_*_read was from a thread that has since been terminated.
This is due to the fact that Windows ties OVERLAPPED I/O operations to
the currently running thread, and if that thread exits, the I/O operation
is aborted.

Signed-off-by: andrei.danila <andrei.danila@analog.com>